# gen-go/models/job.go
package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	strfmt "github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"

	"github.com/go-openapi/errors"
)

// Job job
// swagger:model Job
type Job struct {

	// id
	ID string `json:"id,omitempty"`

	// status
	Status string `json:"status,omitempty"`

	// tasks
	Tasks []*Task `json:"tasks"`

	// workflow
	Workflow *Workflow `json:"workflow,omitempty"`
}

// Validate validates this job
func (m *Job) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateTasks(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if err := m.validateWorkflow(formats); err != nil {
		// prop
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *Job) validateTasks(formats strfmt.Registry) error {

	if swag.IsZero(m.Tasks) { // not required
		return nil
	}

	for i := 0; i < len(m.Tasks); i++ {

		if swag.IsZero(m.Tasks[i]) { // not required
			continue
		}

		if m.Tasks[i] != nil {

			if err := m.Tasks[i].Validate(formats); err != nil {
				return err
			}
		}

	}

	return nil
}

func (m *Job) validateWorkflow(formats strfmt.Registry) error {

	if swag.IsZero(m.Workflow) { // not required
		return nil
	}

	if m.Workflow != nil {

		if err := m.Workflow.Validate(formats); err != nil {
			return err
		}
	}

	return nil
}