# swagger.yml
swagger: '2.0'
info:
  title: workflow-manager
  description: Minimal Workflow orchestrator for AWS Batch
  version: 0.1.0
  x-npm-package: workflow-manager
schemes:
  - http
produces:
  - application/json

responses:
  BadRequest:
    description: "Bad Request"
    schema:
      $ref: "#/definitions/BadRequest"
  InternalError:
    description: "Internal Error"
    schema:
      $ref: "#/definitions/InternalError"
  NotFound:
    description: "Entity Not Found"
    schema:
      $ref: "#/definitions/NotFound"

paths:
  /_health:
    get:
      operationId: healthCheck
      description: Checks if the service is healthy
      responses:
        200:
          description: OK response

  /workflows:
    post:
      operationId: newWorkflow
      summary: Create a new workflow
      parameters:
        - name: NewWorkflowRequest
          in: body
          schema:
            $ref: '#/definitions/NewWorkflowRequest'
      responses:
        201:
          description: Successful creation of a new workflow
          schema:
            $ref: '#/definitions/Workflow'

  /workflows/{name}:
    get:
      summary: Get a Workflow by Name
      operationId: getWorkflowByName
      parameters:
        - name: name
          in: path
          type: string
          required: true
      responses:
        200:
          description: Workflow
          schema:
            $ref: "#/definitions/Workflow"
        404:
          $ref: "#/responses/NotFound"

  /jobs/{workflowName}:
    put:
      summary: Start a job for a workflow
      operationId: startJobForWorkflow
      parameters:
        - name: workflowName
          in: path
          type: string
          required: true
        - name: input
          in: body
          schema:
            $ref: "#/definitions/JobInput"
      responses:
        200:
          description: Respond with Job Id and Name
          schema:
            $ref: "#/definitions/Job"
        404:
          $ref: "#/responses/NotFound"
    get:
      summary: Get summary of all active jobs for workflow
      operationId: getJobsForWorkflow
      parameters:
        - name: workflowName
          in: path
          type: string
          required: true
      responses:
        200:
          description: Job
          schema:
            type: array
            items:
              $ref: "#/definitions/Job"
        404:
          $ref: "#/responses/NotFound"

  /jobs/{workflowName}/{jobId}:
    get:
      summary: Get details for the given jobId
      operationId: GetJob
      parameters:
        - name: workflowName
          in: path
          type: string
          required: true
        - name: jobId
          in: path
          type: string
          required: true
      responses:
        200:
          description: Job
          schema:
            $ref: "#/definitions/Job"
        404:
          $ref: "#/responses/NotFound"

definitions:
  InternalError:
    type: object
    properties:
      message:
        type: string

  BadRequest:
    type: object
    properties:
      message:
        type: string

  NotFound:
    type: object
    properties:
      message:
        type: string

  NewWorkflowRequest:
    type: object
    properties:
      name: 
        type: string
      startAt:
        type: string
      states:
        type: array
        items:
          $ref: '#/definitions/State'

  State:
    type: object
    properties:
      name: 
        type: string
      type: 
        type: string
      comment:
        type: string
      next:
        type: string
      resource:
        type: string
      end:
        type: boolean

  Workflow:
    type: object
    properties:
      id:
        type: string
      revision:
        type: integer
      name: 
        type: string
      startAt:
        type: string
      states:
        type: array
        items:
          $ref: '#/definitions/State'

  Job:
    type: object
    properties:
      id:
        type: string
      workflow:
        $ref: '#/definitions/Workflow'
      status:
        type: string
      tasks:
        type: array
        items:
          $ref: '#/definitions/Task'

  Task:
    type: object
    properties:
      id:
        type: string
      state:
        type: string
      status:
        type: string

  JobInput:
    type: object
    properties:
      data:
        type: string